+++ src/Corporation/Research.ts
@@ -8,6 +8,7 @@ export interface IConstructorParams {
     employeeEffMult?: number;
     employeeIntMult?: number;
     productionMult?: number;
+    productProductionMult?: number;
     salesMult?: number;
     sciResearchMult?: number;
     storageMult?: number;
@@ -30,6 +31,7 @@ export class Research {
     employeeEffMult: number = 1;
     employeeIntMult: number = 1;
     productionMult: number = 1;
+    productProductionMult: number = 1;
     salesMult: number = 1;
     sciResearchMult: number = 1;
     storageMult: number = 1;
@@ -38,14 +40,15 @@ export class Research {
         this.name = p.name;
         this.cost = p.cost;
         this.desc = p.desc;
-        if (p.advertisingMult)    { this.advertisingMult = p.advertisingMult; }
-        if (p.employeeChaMult)    { this.employeeChaMult = p.employeeChaMult; }
-        if (p.employeeCreMult)    { this.employeeCreMult = p.employeeCreMult; }
-        if (p.employeeEffMult)    { this.employeeEffMult = p.employeeEffMult; }
-        if (p.employeeIntMult)    { this.employeeIntMult = p.employeeIntMult; }
-        if (p.productionMult)     { this.productionMult = p.productionMult; }
-        if (p.salesMult)          { this.salesMult = p.salesMult; }
-        if (p.sciResearchMult)    { this.sciResearchMult = p.sciResearchMult; }
-        if (p.storageMult)        { this.storageMult = p.storageMult; }
+        if (p.advertisingMult)          { this.advertisingMult = p.advertisingMult; }
+        if (p.employeeChaMult)          { this.employeeChaMult = p.employeeChaMult; }
+        if (p.employeeCreMult)          { this.employeeCreMult = p.employeeCreMult; }
+        if (p.employeeEffMult)          { this.employeeEffMult = p.employeeEffMult; }
+        if (p.employeeIntMult)          { this.employeeIntMult = p.employeeIntMult; }
+        if (p.productionMult)           { this.productionMult = p.productionMult; }
+        if (p.productProductionMult)    { this.productProductionMult = p.productProductionMult; }
+        if (p.salesMult)                { this.salesMult = p.salesMult; }
+        if (p.sciResearchMult)          { this.sciResearchMult = p.sciResearchMult; }
+        if (p.storageMult)              { this.storageMult = p.storageMult; }
     }
 }