+++ src/NetscriptFunctions.js
@@ -44,6 +44,7 @@ import { AllServers,
 import { Server }                                   from "./Server/Server";
 import { GetServerByHostname,
          getServer,
+         getServerOnNetwork,
          numCycleForGrowth,
          processSingleServerGrowth }                from "./Server/ServerHelpers";
 import { getPurchaseServerCost,